Move MutPtr to its own module
