+++ spatialos-sdk/src/lib.rs
@@ -1,3 +1,5 @@
 extern crate spatialos_sdk_sys;
 
 pub mod worker;
+
+mod ptr;
